Рефакторинг кода. Структурирование проекта.

+++ scr/bot/handlers/message_moderation_handler.py
@@ -1,5 +1,6 @@
 import asyncio
 
+from aiogram.exceptions import TelegramBadRequest
 from aiogram.types import Message
 from loguru import logger
 
@@ -10,7 +11,7 @@ from scr.utils.models import BadWords, get_privileged_users
 
 
 @router.message()
-async def handle_text_messages(message: Message) -> None:
+async def message_moderation_handler(message: Message) -> None:
     """
     Основной обработчик текстовых сообщений. Обрабатывает пересылаемые сообщения, упоминания, запрещенные слова и ссылки.
 
@@ -41,48 +42,50 @@ async def handle_text_messages(message: Message) -> None:
             if (chat_id, user_id) not in privileged_users:
                 await message.delete()
                 warning = await message.answer(
-                    f"<code>✅ {message.from_user.full_name}</code>\n"
-                    f"<code>В чате запрещены пересылаемые сообщения.</code>",
+                    text=translations["ru"]["message_moderation"]["moderation_forward_message"],
                     parse_mode="HTML",
                 )
                 await asyncio.sleep(int(time_del))
                 await warning.delete()
             return
-        # Вывод запрещенных слов с чисткой дублей
-        bad_words = list(set(word.bad_word for word in BadWords.select()))
-        print(bad_words)
-        # Проверка на запрещенные слова
-        for word in bad_words:
-            if word[0].lower() in message.text.lower():
-                # recording_actions_in_the_database(word[0], message)
-                await message.delete()
-                warning = await message.answer(
-                    f"⚠ В вашем сообщении обнаружено запрещенное слово: <code>{word[0]}</code>. "
-                    f"Пожалуйста, не используйте его в дальнейшем.",
-                    parse_mode="HTML",
-                )
-                await asyncio.sleep(int(time_del))
-                await warning.delete()
-                return  # После удаления сообщения больше ничего не проверяем
 
-        # Проверка на ссылки
-        for entity in message.entities or []:
-            if entity.type in ["url", "text_link", "mention"]:
-                privileged_users = get_privileged_users()
-                if (chat_id, user_id) not in privileged_users:
-                    await message.delete()
+        if message.text:
+            # Вывод запрещенных слов с чисткой дублей
+            bad_words = list(set(word.bad_word for word in BadWords.select()))
+            print(bad_words)
+            # Проверка на запрещенные слова
+            for word in bad_words:
+                if word[0].lower() in message.text.lower():
+                    try:
+                        await message.delete()
+                    except TelegramBadRequest:
+                        logger.error("Ошибка при удалении сообщения, так как оно уже удалено ранее")
                     warning = await message.answer(
-                        f"<code>✅ {message.from_user.full_name}</code>\n"
-                        f"<code>В чате запрещена публикация сообщений со ссылками.</code>",
+                        text=translations["ru"]["message_moderation"]["moderation_bad_words"],
                         parse_mode="HTML",
                     )
                     await asyncio.sleep(int(time_del))
                     await warning.delete()
                     return  # После удаления сообщения больше ничего не проверяем
 
+        # Проверка на ссылки
+        if message.text and message.entities:
+            for entity in message.entities:
+                if entity.type in ["url", "text_link", "mention"]:
+                    privileged_users = get_privileged_users()
+                    if (chat_id, user_id) not in privileged_users:
+                        await message.delete()
+                        warning = await message.answer(
+                            text=translations["ru"]["message_moderation"]["moderation_url_message"],
+                            parse_mode="HTML",
+                        )
+                        await asyncio.sleep(int(time_del))
+                        await warning.delete()
+                        return  # После удаления сообщения больше ничего не проверяем
+
     except Exception as e:
-        logger.error(f"Ошибка при обработке текстового сообщения: {e}")
+        logger.exception(f"{e}")
 
 
 def register_text_messages_handlers() -> None:
-    router.message.register(handle_text_messages)
+    router.message.register(message_moderation_handler)
